- edit base url - edit transaction model (Server, Repo, Ui)

+++ app/src/main/java/com/example/mvvm_architechture/domain/repomodels/TransactionRepoModel.java
@@ -7,7 +7,7 @@ public class TransactionRepoModel {
     private Integer tranId = -1;
     private int type = -1;
     private Integer unreadChatCount = -1;
-    private Integer updateTime = -1;
+    private Long updateTime = -1L;
     private String firstName = "";
     private String lastName = "";
     private Integer userId = -1;
@@ -41,10 +41,10 @@ public class TransactionRepoModel {
     public void setUnreadChatCount(Integer unreadChatCount) {
         this.unreadChatCount = unreadChatCount;
     }
-    public Integer getUpdateTime() {
+    public Long getUpdateTime() {
         return updateTime;
     }
-    public void setUpdateTime(Integer updateTime) {
+    public void setUpdateTime(Long updateTime) {
         this.updateTime = updateTime;
     }
     public String getFirstName() {
