+++ app/src/main/java/com/example/mvvm_architechture/domain/repomodels/TransactionsConversationRepoModel.java
@@ -1,11 +1,10 @@
 package com.example.mvvm_architechture.domain.repomodels;
 
-import com.example.mvvm_architechture.domain.servermodels.TransactionServerModel;
 import java.util.List;
 
 public class TransactionsConversationRepoModel {
     private Double balance;
-    private Integer lastGetTime;
+    private Long lastGetTime;
     private String responseMessage;
     private String responseValue;
     private List<TransactionRepoModel> transactions = null;
@@ -18,11 +17,11 @@ public class TransactionsConversationRepoModel {
         this.balance = balance;
     }
 
-    public Integer getLastGetTime() {
+    public Long getLastGetTime() {
         return lastGetTime;
     }
 
-    public void setLastGetTime(Integer lastGetTime) {
+    public void setLastGetTime(Long lastGetTime) {
         this.lastGetTime = lastGetTime;
     }
